requirements.txt
Flask==3.0.2
mysql-connector-python==8.3.0
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1  
Jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2